refactor(multiverse): Move the timeline rendering into a separate widget
